# marketing-os/packages/api-contract/openapi.yaml
...
openapi: 3.1.0
info:
  title: 'Marketing Intelligence OS API'
  version: 1.0.0-phase1
  description: 'Cookie-session API (Sanctum-style CSRF flow: GET /sanctum/csrf-cookie, then send X-XSRF-TOKEN). Every /orgs/{org}/* route is tenant-scoped; non-members receive 404. Errors are RFC 9457 problem details. No endpoint ever returns fabricated metrics: see MetricEnvelope.status.'
servers:
  -
    url: /api/v1
security:
  -
    cookieAuth: []
paths:
...
  '/me/sessions/{id}':
    delete:
      summary: 'Revoke a device session'
      tags:
        - Me
      responses:
        204:
          description: Revoked
        401:
          $ref: '#/components/responses/Unauthenticated'
        404:
          $ref: '#/components/responses/NotFound'
      operationId: revokeSession
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mios_session
...
  responses:
    Unauthenticated:
      description: 'Not signed in (code: unauthenticated)'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: 'Not found, or not a member of the organization (no existence leak)'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        request_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        action:
          type: object
          properties:
            label:
              type: string
            href:
              type: string
        retry_after:
          type: integer
        feature:
          type: string
      required:
        - type
        - title
        - status
        - code
...